Split technologies and language types into 2 files - Makes it easier to manage
