+++ components/Modal/SkillsModal.tsx
@@ -1,10 +1,9 @@
-import { Language, Skill } from "@/types/languagesSkillsTechnologies";
+import { Language, Skill } from "@/types/languages";
 import React from "react";
 import Tag from "../Atoms/Tag";
 import HeadingThree from "../Content/Text/HeadingThree";
-import HeadingTwo from "../Content/Text/HeadingTwo";
-import Modal from "./Modal";
 import Dropdown from "../DropDown/DropDownMenu";
+import Modal from "./Modal";
 
 interface SkillsModalProps {
   languages: Language[];
