+++ types/languages.ts
@@ -7,7 +7,7 @@ interface Skill {
     | "Frontend Web Development"
     | "Backend Web Development"
     | "Full Stack Web Development"
-    | "Object Relation Mapper"
+    | "Object Relational Mapper"
     | "Testing"
     | "Project Manager"
     | "Machine Learning"
@@ -31,22 +31,7 @@ interface Language {
   repositories?: Repository[];
 }
 
-interface Technology {
-  name: string;
-  category?:
-    | "Version Control System"
-    | "Continuous Integration"
-    | "Database"
-    | "Containerization"
-    | "Web Development"
-    | "API"
-    | "Backend Web Development"
-    | "Frontend Web Development"
-    | "Web Socket"
-    | "Other";
-}
-
-export type { Skill, Repository, Language, Technology };
+export type { Skill, Repository, Language };
 
 /**
  * Array of skills for JavaScript and TypeScript.
@@ -60,9 +45,9 @@ const JavaScriptTypeScriptSkills: Skill[] = [
   { skill: "Vitest", category: "Testing" },
   { skill: "Chakra UI", category: "Frontend Web Development" },
   { skill: "Next UI", category: "Frontend Web Development" },
-  { skill: "Prisma", category: "Object Relation Mapper" },
-  { skill: "Drizzle", category: "Object Relation Mapper" },
-  { skill: "Mongoose", category: "Object Relation Mapper" },
+  { skill: "Prisma", category: "Object Relational Mapper" },
+  { skill: "Drizzle", category: "Object Relational Mapper" },
+  { skill: "Mongoose", category: "Object Relational Mapper" },
 ];
 
 /**
@@ -79,7 +64,7 @@ const languages: Language[] = [
       { skill: "Numpy", category: "Machine Learning" },
       { skill: "Jupyter Notebooks", category: "Machine Learning" },
       { skill: "PyTest", category: "Testing" },
-      { skill: "SQLAlchemy", category: "Object Relation Mapper" },
+      { skill: "SQLAlchemy", category: "Object Relational Mapper" },
     ],
     repositories: [
       {
@@ -134,26 +119,4 @@ const languages: Language[] = [
   },
 ];
 
-/**
- * Array of technologies.
- */
-const technologies: Technology[] = [
-  { name: "Git", category: "Version Control System" },
-  { name: "SVN", category: "Version Control System" },
-  { name: "Firebase", category: "Backend Web Development" },
-  { name: "Supabase", category: "Backend Web Development" },
-  { name: "Tailwind CSS", category: "Frontend Web Development" },
-  { name: "Radix UI", category: "Frontend Web Development" },
-  { name: "Headless UI", category: "Frontend Web Development" },
-  { name: "GitHub Actions", category: "Continuous Integration" },
-  { name: "Jenkins", category: "Continuous Integration" },
-  { name: "Docker", category: "Containerization" },
-  { name: "REST", category: "API" },
-  { name: "GraphQL", category: "API" },
-  { name: "PostgreSQL", category: "Database" },
-  { name: "MongoDB", category: "Database" },
-  { name: "Socket.IO", category: "Web Socket" },
-  { name: "Pusher", category: "Web Socket" },
-];
-
-export { languages, technologies };
+export { languages };
